requirements.txt
PyTorch == 2.0.1
transformers == 4.23.1
diffusers == 0.11.1